Add snapshot-with-retention script.
RightScale provides a script for backing up our database's EBS volumes;
however, it is very specifically designed for that, and makes a bunch of
assumptions that mean we can't use it for other machines.

There are a lot of people who have written scripts similar to this, but
they generally a) require big SDKs or b) don't work.  This one mostly
worked, so I fixed it and refactored a bunch, and now it's ok.

It's very simple, and I want more complex behavior - namely, parameters
to control the cutoff date and several combined cutoff dates (keep 7
daily snapshots, 4 weekly snapshots, and 6 monthly).  This will work for
now, though.