Migrate snapshot-with-retention to aws-cli.
The normal AWS cli tools are being a pain, mostly due to being written
in Java.  Now we use the (also official) ones written in Python, which
are much easier to get working on a system.

Since these tools are meant to take in and put out JSON, some of the
invocations are a bit more awkward than they were previously.  I'm
willing to pay this price to have something that works.